backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
python-dotenv==1.0.1
httpx==0.27.2
pymongo[srv]==4.9.1
motor==3.6.0
